Restructured tests.
